Add files via upload
This program scans a series of pdb files to find instances where the 2' and 3' OH of a nucleotide H-bond with 2 consecutive backbone NH of a peptide. The algorithm searches for O2' and O3' atoms in individual residues, stores those in two arrays, and also searches for all backbone NHs and stores those in an array. Next, the O3' atoms are scanned for H-bonding with any NHs. When a hit is found, the O2' atoms are scanned for atoms in the same residue as the O3' atom which also H bond with the NH in the next residue from the NH previously identified. When a hit is found, lines are printed for a .pml file to load the pdb file, align the O2' residue on the 3' terminus of RocR, and show lines for the nucleotide, as well as the 2 consecutive amino acids in H-bonding distance.