Add files via upload
This program scans a series of pdb files to find instances where the N-terminus of a helix interacts with two consecutive phosphate groups in an RNA chain. The helix  must be capped with a Ser, whose OG H-bonds to the maintain N and also to one of the phosphates. The N atoms of the the two residues C-terminal to the Ser H-bond to the phosphate 3' to the one that is recognized by the Ser. This arrangement is found in in the RocC-RocR structure and we think it is conserved in the FinO and ProQ mechanisms.
       The algorithm searches for O1P, O2P, backbone N, and Ser OG and stores those in 4 arrays. Next, it loops through these arrays and makes new arrays for OG-OP1, OG-N, OP1-N, and OP2-N H bonds where it stores the chain and resids of the bonded atoms. Finally, it loops through the bond arrays to find instances where all the H bonds are made.